Refactor validate-on-value Atomic tests to own module, to match other validate-on-value tests
